# devintel-backend/app/schemas/chat.py
"""Chat schemas."""

from datetime import datetime
from typing import Optional
from uuid import UUID

from pydantic import BaseModel, Field


class ChatHistoryMessage(BaseModel):
    """A single message in chat history."""

    role: str = Field(..., pattern="^(user|assistant)$", description="Message role")
    content: str = Field(..., min_length=1, max_length=5000, description="Message content")


class ChatRequest(BaseModel):
    """Chat request schema."""

    repository_id: UUID = Field(..., description="Repository ID to query")
    question: str = Field(..., min_length=1, max_length=2000, description="User question")
    chat_history: list[ChatHistoryMessage] = Field(
        default_factory=list,
        max_length=20,
        description="Previous chat messages for multi-turn context (max 20)",
    )
    agent_type: Optional[str] = Field(
        None,
        description="Optional agent override: architect, security, performance, test_engineer",
    )


class ChatResponse(BaseModel):
    """Chat response schema."""

    id: UUID
    repository_id: UUID
    question: str
    response: str
    token_usage: int
    response_time_ms: Optional[int] = None
    agent_type: Optional[str] = None
    created_at: datetime

    class Config:
        """Pydantic config."""

        from_attributes = True


class ChatHistoryRecord(BaseModel):
    """Record retrieved from database."""

    role: str
    content: str
    timestamp: datetime


class ChatHistoryResponse(BaseModel):
    """Chat history response schema."""

    messages: list[ChatHistoryRecord]
    repository_id: UUID


class StreamChunk(BaseModel):
    """Stream chunk schema for SSE."""

    content: str
    done: bool = False


class AgentDraftRequest(BaseModel):
    """Request schema for an autonomous agent action draft."""

    repository_id: UUID = Field(..., description="Repository ID to act upon")
    instruction: str = Field(..., min_length=10, max_length=2000, description="The instruction for the agent (e.g., 'Refactor button component')")


class FileChange(BaseModel):
    """A file change proposal from the agent."""

    path: str
    content: str


class DraftPayload(BaseModel):
    """The structured draft proposal generated by the LLM."""

    branch_name: str
    pr_title: str
    pr_body: str
    commit_message: str
    file_changes: list[FileChange]
    generate_tests: Optional[bool] = False


class AgentDraftResponse(BaseModel):
    """Response schema giving the drafted proposal to the user."""

    draft: DraftPayload = Field(..., description="The planned changes to review")


class AgentExecuteRequest(BaseModel):
    """Request schema to execute an approved draft."""

    repository_id: UUID
    draft: DraftPayload


class AgentExecuteResponse(BaseModel):
    """Response schema upon successful PR creation."""

    pr_url: str = Field(..., description="The URL of the generated Pull Request")
    pr_number: int = Field(..., description="The PR number")
    branch_name: str = Field(..., description="The name of the generated branch")


class AgentExecuteWithTestsResponse(BaseModel):
    """Response schema for PR execution with test status."""

    status: str = Field(..., description="Execution status: success or tests_failed")
    pr_url: Optional[str] = None
    pr_number: Optional[int] = None
    branch_name: Optional[str] = None
    test_id: Optional[str] = None
    output: Optional[str] = None
    message: Optional[str] = None