# devintel-backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DevIntel AI",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/chat/{repository_id}/stream": {
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "Stream Chat",
        "description": "Stream an AI chat response for a repository using RAG.\n\n    Returns a Server-Sent Events (SSE) stream.\n    Each event is: `data: <json>\n\n`\n    where json is one of:\n      - `{ \"type\": \"chunk\", \"content\": \"...\" }`\n      - `{ \"type\": \"done\" }`\n      - `{ \"type\": \"error\", \"message\": \"...\" }`",
        "operationId": "stream_chat_api_v1_chat__repository_id__stream_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "repository_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Repository Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat/draft": {
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "Agent Draft",
        "description": "Generate a draft PR proposal for user review.",
        "operationId": "agent_draft_api_v1_chat_draft_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentDraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDraftResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/chat/execute": {
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "Agent Execute",
        "description": "Execute an approved draft PR on GitHub. Optionally generates tests first.",
        "operationId": "agent_execute_api_v1_chat_execute_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentExecuteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/AgentExecuteResponse"
                    },
                    {
                      "$ref": "#/components/schemas/AgentExecuteWithTestsResponse"
                    }
                  ],
                  "title": "Response Agent Execute Api V1 Chat Execute Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/collab/sessions": {
      "post": {
        "tags": [
          "Collaboration"
        ],
        "summary": "Create Collaboration Session",
        "description": "Create a new collaboration session.",
        "operationId": "create_collaboration_session_api_v1_collab_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollaborationSessionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollaborationSessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/collab/sessions/{repository_id}": {
      "get": {
        "tags": [
          "Collaboration"
        ],
        "summary": "Get Active Session",
        "description": "Get active collaboration session for a repository.",
        "operationId": "get_active_session_api_v1_collab_sessions__repository_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "repository_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Repository Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollaborationSessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/collab/sessions/{session_id}/history": {
      "get": {
        "tags": [
          "Collaboration"
        ],
        "summary": "Get Session History",
        "description": "Get message history for a collaboration session.",
        "operationId": "get_session_history_api_v1_collab_sessions__session_id__history_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollaborationHistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentDraftRequest": {
        "properties": {
          "repository_id": {
            "type": "string",
            "format": "uuid",
            "title": "Repository Id",
            "description": "Repository ID to act upon"
          },
          "instruction": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 10,
            "title": "Instruction",
            "description": "The instruction for the agent (e.g., 'Refactor button component')"
          }
        },
        "type": "object",
        "required": [
          "repository_id",
          "instruction"
        ],
        "title": "AgentDraftRequest",
        "description": "Request schema for an autonomous agent action draft."
      },
      "AgentDraftResponse": {
        "properties": {
          "draft": {
            "$ref": "#/components/schemas/DraftPayload",
            "description": "The planned changes to review"
          }
        },
        "type": "object",
        "required": [
          "draft"
        ],
        "title": "AgentDraftResponse",
        "description": "Response schema giving the drafted proposal to the user."
      },
      "AgentExecuteRequest": {
        "properties": {
          "repository_id": {
            "type": "string",
            "format": "uuid",
            "title": "Repository Id"
          },
          "draft": {
            "$ref": "#/components/schemas/DraftPayload"
          }
        },
        "type": "object",
        "required": [
          "repository_id",
          "draft"
        ],
        "title": "AgentExecuteRequest",
        "description": "Request schema to execute an approved draft."
      },
      "AgentExecuteResponse": {
        "properties": {
          "pr_url": {
            "type": "string",
            "title": "Pr Url",
            "description": "The URL of the generated Pull Request"
          },
          "pr_number": {
            "type": "integer",
            "title": "Pr Number",
            "description": "The PR number"
          },
          "branch_name": {
            "type": "string",
            "title": "Branch Name",
            "description": "The name of the generated branch"
          }
        },
        "type": "object",
        "required": [
          "pr_url",
          "pr_number",
          "branch_name"
        ],
        "title": "AgentExecuteResponse",
        "description": "Response schema upon successful PR creation."
      },
      "AgentExecuteWithTestsResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Execution status: success or tests_failed"
          },
          "pr_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pr Url"
          },
          "pr_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pr Number"
          },
          "branch_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Branch Name"
          },
          "test_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Test Id"
          },
          "output": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "AgentExecuteWithTestsResponse",
        "description": "Response schema for PR execution with test status."
      },
...
      "ChatHistoryMessage": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role"
          },
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "role",
          "content"
        ],
        "title": "ChatHistoryMessage"
      },
      "ChatRequest": {
        "properties": {
          "question": {
            "type": "string",
            "title": "Question"
          },
          "chat_history": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ChatHistoryMessage"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chat History"
          }
        },
        "type": "object",
        "required": [
          "question"
        ],
        "title": "ChatRequest"
      },
      "CollaborationHistoryResponse": {
        "properties": {
          "messages": {
            "items": {
              "$ref": "#/components/schemas/CollaborationMessageResponse"
            },
            "type": "array",
            "title": "Messages"
          }
        },
        "type": "object",
        "required": [
          "messages"
        ],
        "title": "CollaborationHistoryResponse",
        "description": "Collaboration message history response."
      },
      "CollaborationMessageResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "message_type": {
            "type": "string",
            "title": "Message Type"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "file_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Path"
          },
          "cursor_line": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cursor Line"
          },
          "cursor_column": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cursor Column"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "session_id",
          "user_id",
          "message_type",
          "content",
          "file_path",
          "cursor_line",
          "cursor_column",
          "created_at"
        ],
        "title": "CollaborationMessageResponse",
        "description": "Collaboration message response."
      },
      "CollaborationSessionCreate": {
        "properties": {
          "repository_id": {
            "type": "string",
            "format": "uuid",
            "title": "Repository Id",
            "description": "Repository to collaborate on"
          },
          "session_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Session Name"
          }
        },
        "type": "object",
        "required": [
          "repository_id",
          "session_name"
        ],
        "title": "CollaborationSessionCreate",
        "description": "Create collaboration session request."
      },
      "CollaborationSessionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "repo_id": {
            "type": "string",
            "format": "uuid",
            "title": "Repo Id"
          },
          "owner_id": {
            "type": "string",
            "format": "uuid",
            "title": "Owner Id"
          },
          "session_name": {
            "type": "string",
            "title": "Session Name"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "participants_count": {
            "type": "integer",
            "title": "Participants Count"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "repo_id",
          "owner_id",
          "session_name",
          "is_active",
          "participants_count",
          "created_at",
          "updated_at"
        ],
        "title": "CollaborationSessionResponse",
        "description": "Collaboration session response."
      },
...
      "DraftPayload": {
        "properties": {
          "branch_name": {
            "type": "string",
            "title": "Branch Name"
          },
          "pr_title": {
            "type": "string",
            "title": "Pr Title"
          },
          "pr_body": {
            "type": "string",
            "title": "Pr Body"
          },
          "commit_message": {
            "type": "string",
            "title": "Commit Message"
          },
          "file_changes": {
            "items": {
              "$ref": "#/components/schemas/FileChange"
            },
            "type": "array",
            "title": "File Changes"
          },
          "generate_tests": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generate Tests",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "branch_name",
          "pr_title",
          "pr_body",
          "commit_message",
          "file_changes"
        ],
        "title": "DraftPayload",
        "description": "The structured draft proposal generated by the LLM."
      },
...
      "FileChange": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "path",
          "content"
        ],
        "title": "FileChange",
        "description": "A file change proposal from the agent."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}